Cleaning up threading in HttpClient and PollSet
Introduced usage of thread safety analysis from CLang. It helped
tracking some of potential issues in HttpClient, but created a lot of
side problems with usage of other libc++.

Fixed typo with setting MINIROS_COMPILE_OPTIONS. Now required level of
warnings is properly set.

Moved all small helpers related to threading into internal/threading.h
header. Not it also defines all markup macros required by thread safety
analysis from CLang.

Added flag to PollSet to start without internal signaling.

Fixed some of compiler warnings.

+++ src/http/http_tools.cpp
@@ -91,7 +91,7 @@ void HttpParserFrame::finish(bool request)
     size_t currentLength = m_bodyPosition;
     if (m_contentLength >= 0)
       currentLength += m_contentLength;
-    if (m_contentLength  == data.size()) {
+    if (static_cast<size_t>(m_contentLength)  == data.size()) {
       data.clear();
     } else {
       // Truncate all data related to current request
@@ -359,7 +359,7 @@ int HttpParserFrame::incrementalParse()
   if (m_state == HttpParserFrame::ParseBody) {
     if (m_contentLength >= 0) {
       // Need body and got all the data.
-      if (m_bodyPosition + m_contentLength <= data.length()) {
+      if (static_cast<size_t>(m_bodyPosition + m_contentLength) <= data.length()) {
         m_state = HttpParserFrame::ParseComplete;
         m_bodyEnd = m_bodyPosition + m_contentLength;
         cp = start + m_bodyEnd;
